refactor: extract patsing to _internal module
