+++ src/conformly/_internal/parsing/adapters/registry.py
@@ -1,7 +1,5 @@
 from .protocol import ParcingAdapterProtocol
 
-from conformly.specs import ModelSpec
-
 _adapters: list[ParcingAdapterProtocol] = []
 
 
@@ -14,7 +12,3 @@ def get_adapter(model: type) -> ParcingAdapterProtocol:
         if adapter.supports(model):
             return adapter
     raise TypeError(f"No adapters found for {model!r}")
-
-
-def parse_model(model: type) -> ModelSpec:
-    return get_adapter(model).parse(model)
