+++ src/conformly/_internal/parsing/extractors/constraints.py
@@ -1,7 +1,7 @@
 from collections.abc import Mapping
 from typing import Annotated, Any, cast, get_args, get_origin
 
-from ..constraints import (
+from ....constraints import (
     ALLOWED_CONSTRAINT_TYPE,
     NUMERIC_CONSTRAINTS,
     STRING_CONSTRAINTS,
@@ -9,7 +9,7 @@ from ..constraints import (
     ConstraintType,
     OneOf,
 )
-from ..constraints.mapping import create_constraint
+from ....constraints.mapping import create_constraint
 
 
 def is_constraints_consistent(constraints: tuple[Constraint, ...]) -> bool:
