--- src/conformly/parsing/__init__.py
@@ -1,4 +1,0 @@
-from .protocol import ParcingAdapterProtocol
-from .registry import get_adapter, parse_model, register
-
-__all__ = ["ParcingAdapterProtocol", "get_adapter", "parse_model", "register"]
