--- src/conformly/specs/__init__.py
@@ -1,4 +1,0 @@
-from .field import FieldSpec
-from .model import ModelSpec
-
-__all__ = ["FieldSpec", "ModelSpec"]
