+++ tests/benchmarks/test_dataclass.py
@@ -4,6 +4,7 @@ from typing import Annotated, Literal
 import pytest
 
 from conformly import case, cases
+from conformly._internal.parsing.adapters.dataclass import parse
 from conformly.constraints import (
     GreaterOrEqual,
     GreaterThan,
@@ -12,7 +13,6 @@ from conformly.constraints import (
     MinLength,
     Pattern,
 )
-from conformly.parsing.adapters.dataclass import parse
 from conformly.resolver.resolve import resolve_model
