+++ tests/benchmarks/test_pydantic.py
@@ -6,7 +6,7 @@ from typing import Literal
 from pydantic import BaseModel, Field
 
 from conformly import case, cases
-from conformly.parsing.adapters.pydantic import parse
+from conformly._internal.parsing.adapters.pydantic import parse
 from conformly.resolver.resolve import resolve_model
