+++ tests/test_parsing/test_constraints_extractors.py
@@ -3,16 +3,16 @@ from typing import Annotated
 
 import pytest
 
-from conformly.constraints import Constraint
-from conformly.constraints.enum import OneOf
-from conformly.constraints.numeric import GreaterOrEqual
-from conformly.constraints.string import MinLength, Pattern
-from conformly.parsing.constraints import (
+from conformly._internal.parsing.extractors.constraints import (
     _metadata_to_constraints,
     is_constraints_consistent,
     parse_annotated_constraints,
     parse_metadata_constraints,
 )
+from conformly.constraints import Constraint
+from conformly.constraints.enum import OneOf
+from conformly.constraints.numeric import GreaterOrEqual
+from conformly.constraints.string import MinLength, Pattern
 
 
 @dataclass
