+++ tests/test_parsing/test_pydantic_adapter.py
@@ -12,6 +12,15 @@ from pydantic.fields import FieldInfo
 from pydantic.types import StringConstraints
 from pydantic_core import PydanticUndefined
 
+from conformly._internal.parsing import ModelSpec
+from conformly._internal.parsing.adapters.pydantic import (
+    _parse_default,
+    _parse_fieldinfo_constraints,
+    parse,
+    parse_field,
+    parse_fields,
+    supports,
+)
 from conformly.constraints import (
     GreaterOrEqual,
     GreaterThan,
@@ -23,15 +32,6 @@ from conformly.constraints import (
     Pattern,
 )
 from conformly.fields import Email
-from conformly.parsing.adapters.pydantic import (
-    _parse_default,
-    _parse_fieldinfo_constraints,
-    parse,
-    parse_field,
-    parse_fields,
-    supports,
-)
-from conformly.specs import ModelSpec
 from conformly.types import _UNSET, ENUMERATED_TYPE
